Agregar interfaces IBodegaRepository, IIngresoActivoRepository y IPaqueteActivoRepository

+++ src/main/java/com/example/bodegaesfeagape/Entidades_de_Negocio/IngresoActivo.java
@@ -11,7 +11,7 @@ public class IngresoActivo {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne
     @JoinColumn(name = "usuario_id")
@@ -58,12 +58,14 @@ public class IngresoActivo {
         this.usuarioId = usuarioId;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
+
+   
 
 }
